Create database connection
A PDO wrapper class was created to allow the app to interact with
a database.

The current implementation initializes a local sqlite database,
creates tables using up to date schemas, and inserts demo data
from a json file.

The project was renamed to oversight, which required changes in
the namespaces.