Write contributor class
A contributor and contributor collection model was created in order
to provide structure to the data fetched from the github API. Prior
to the refactor data fetched from was directly inserted into the
database.

Using a class provides a standard structure for the contributor
data than can be used throughout the code base.