Integrate Slim and Twig
A configuration directory was created to configure middleware and
routes with Slim.

The index.php file was refactored to load bootstrap script that
builds a container using PHP-DI; creates the app; and sets up
routes and middleware.

The admin.php file was removed and the logic is now handled
by the AdminController, in order to follow MVC.

The single Controller class was removed to be replaced with individual
controllers to improve modularity. The controllers are to be
refactored so that they fetch data from the database and not
from the github REST API. Only the AdminController will
update the database through the github REST API.

The Slim and Twig integration was based on the following
tutorials:
https://odan.github.io/2019/11/05/slim4-tutorial.html
https://odan.github.io/2020/04/17/slim4-twig-templates.html