Refactor PluginsController
The PluginsController was refactored to fetch data from the
database.

The custom twig filter and helper functions were removed, because
PDO returns an array and so the custom twig filter and helper functions
were no longer needed in order to convert a class object into an array.

A method was added to the DB class to query the database for all
records of a table.