Extend data input
Data fields where added to the plugins table and the data presented
to the use was also extended.

In the Plugin class the method was refactored to display issues by
age only if they are open. There is still a bug that prevents the
oldest issue from being fetched.

Bootstrap was added to the twig templates and a base template
was created.