Writing method to get report data
Created a issues model and table schema, in order to add issues
data to the database neede for the report. The issues will be inserted
into the database automatically without any url parameters in the
routes. The routes for updating need refactoring to follow a
common pattern and reduce latency.

The admin controller needs refactoring, because it is becoming
bloated and hard to follow.